eclipse commit test

+++ server/game/test/org/tnt/test/TestSimulator.java
@@ -20,7 +20,7 @@ public class TestSimulator extends GameSimulator
 	}
 
 	@Override
-	public void step( long stepTime, long gameTime )
+	public void step( long stepTime, int gameTime )
 	{
 	}
